package/systemd: fix build with old toolchains
Toolchains using glibc-2.18 or older do not define O_TMPFILE, which
causes build failures on some archs.

systemd has a definition for O_TMPFILE if it is missing, but only
defines it for i386 or x86_64. Furthermore, the header defining it is
not included everywhere O_TMPFILE is used.

Fix that with three patches backported from upstream:
  - include the needed header where it is needed (he!),
  - define O_TMPFILE for all archs, according to linux-4.8rc3,
  - no longer guard against undefined O_TMPFILE in fileio.

Upstream merge commit:
    https://github.com/systemd/systemd/commit/4a13100c6a5a0a4b793e90bd43d21c3696c42d46

Fixes:

  http://autobuild.buildroot.net/results/b0067e72ffcbbe1db9ef49ab297cece951345aeb/

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>